Work around `gh-pages`' wrong linking to root
Relative links are resolving with `joaojgabriel.github.io/`as a base
URL (e.g. `images/foo-bar.webp` would reference
`joaojgabriel.github.io/images/foo-bar.webp`), thus returning a
`404 NOT FOUND` response, since that is not where `images/` is located.

Running local files presents the same issue with root-relative links,
as explained in internetingishard's [links and images tutorial](https://www.internetingishard.com/html-and-css/links-and-images/#absolute-relative-and-root-relative-links) (albeit
with different specific paths: referring to `/images/foo-bar.webp` in
`/recipes/foo-bar.html` will take us to the root directory of the
operating system, instead of the root of the project at hand).
Root-relative links are only supported in servers (and that includes
`http://localhost:[port]/`).

I've found two solutions to this problem:
 - Use Jekyll pre-deployment, which lets you preview it in a local
   server (i.e. a `localhost`), and change the base URL manually (as
   explained in this [stackoverflow answer](https://stackoverflow.com/a/19173888/17743255))
 - Conveniently place directories, and use relative links (instead of
   root-relative ones)

The first way might improve developer experience and come in handy for
larger projects, but it is too cumbersome to be justified in this one,
so go with the second and call it a day.